+++ R/AllClasses.R
@@ -221,11 +221,15 @@ CrunchDataset <- setClass("CrunchDataset",
     contains = c("ShojiObject"),
     slots = c(
         variables = "VariableCatalog",
+        hiddenVariables = "VariableCatalog",
+        privateVariables = "VariableCatalog",
         filter = "CrunchLogicalExpr",
         tuple = "DatasetTuple"
     ),
     prototype = prototype(
         variables = VariableCatalog(),
+        hiddenVariables = VariableCatalog(),
+        privateVariables = VariableCatalog(),
         filter = CrunchLogicalExpr(),
         tuple = DatasetTuple()
     )