+++ vignettes/fork-and-merge/3/api/sources-POST.R
@@ -1,20 +1,20 @@
 structure(list(url = "/api/sources/", status_code = 201L, headers = structure(list(
-    allow = "GET, HEAD, OPTIONS, POST", `content-encoding` = "gzip", 
-    `content-type` = "application/json;charset=utf-8", date = "Mon, 14 May 2018 16:51:33 GMT", 
-    location = "/api/sources/1d2e8d/", server = "nginx", `set-cookie` = "REDACTED", 
-    vary = "Cookie, Accept-Encoding", `x-timing` = "", `content-length` = "86", 
-    connection = "keep-alive"), class = c("insensitive", "list"
-)), all_headers = list(list(status = 201L, version = "HTTP/1.1", 
-    headers = structure(list(allow = "GET, HEAD, OPTIONS, POST", 
-        `content-encoding` = "gzip", `content-type` = "application/json;charset=utf-8", 
-        date = "Mon, 14 May 2018 16:51:33 GMT", location = "/api/sources/1d2e8d/", 
-        server = "nginx", `set-cookie` = "REDACTED", vary = "Cookie, Accept-Encoding", 
-        `x-timing` = "", `content-length` = "86", connection = "keep-alive"), class = c("insensitive", 
+    date = "Mon, 04 May 2020 23:13:22 GMT", `content-type` = "application/json;charset=utf-8", 
+    `content-length` = "86", location = "/api/sources/8327ed/", 
+    server = "nginx", `content-encoding` = "gzip", vary = "Cookie, Accept-Encoding", 
+    allow = "GET, HEAD, OPTIONS, POST", `x-timing` = "", `set-cookie` = "REDACTED"), class = c("insensitive", 
+"list")), all_headers = list(list(status = 201L, version = "HTTP/2", 
+    headers = structure(list(date = "Mon, 04 May 2020 23:13:22 GMT", 
+        `content-type` = "application/json;charset=utf-8", `content-length` = "86", 
+        location = "/api/sources/8327ed/", server = "nginx", 
+        `content-encoding` = "gzip", vary = "Cookie, Accept-Encoding", 
+        allow = "GET, HEAD, OPTIONS, POST", `x-timing` = "", 
+        `set-cookie` = "REDACTED"), class = c("insensitive", 
     "list")))), cookies = structure(list(domain = ".crunch.io", 
-    flag = TRUE, path = "/", secure = FALSE, expiration = structure(1557852694, class = c("POSIXct", 
+    flag = TRUE, path = "/", secure = FALSE, expiration = structure(1620170002, class = c("POSIXct", 
     "POSIXt")), name = "token", value = "REDACTED"), row.names = c(NA, 
 -1L), class = "data.frame"), content = charToRaw("{\"element\": \"shoji:view\", \"self\": \"/api/sources/\"}"), 
-    date = structure(1526316693, class = c("POSIXct", "POSIXt"
-    ), tzone = "GMT"), times = c(redirect = 0, namelookup = 2.8e-05, 
-    connect = 3e-05, pretransfer = 0.000153, starttransfer = 0.000166, 
-    total = 1.68213)), class = "response")
+    date = structure(1588634002, class = c("POSIXct", "POSIXt"
+    ), tzone = "GMT"), times = c(redirect = 0, namelookup = 2.4e-05, 
+    connect = 2.5e-05, pretransfer = 0.000142, starttransfer = 0.000147, 
+    total = 0.88913)), class = "response")